ml-services/requirements.txt
requests==2.31.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
numpy>=1.21.0,<2.0.0
scikit-learn==1.5.2
pytesseract==0.3.13
Pillow==10.0.0
pdfplumber==0.9.0
pypdf==4.0.0
pdf2image==1.16.3
sentence-transformers==2.2.2
python-multipart==0.0.6
# faiss-cpu 1.7.4 is no longer available on PyPI for many platforms/Python versions
# Use a newer faiss-cpu available on PyPI (and/or use conda-forge on Windows).
faiss-cpu==1.13.0
transformers==4.38.0
torch==2.2.2
tqdm==4.66.1
beautifulsoup4==4.12.3
lxml==5.2.2
selenium==4.21.0
webdriver-manager==4.0.1

